dvgl_benchmark/requirements.txt
numpy==1.19.4
torchvision==0.8.1
psutil==5.6.7
faiss_cpu==1.5.3
tqdm==4.48.2
torch==1.7.0
Pillow==8.2.0
scikit_learn==0.24.1
torchscan==0.1.1
googledrivedownloader==0.4
requests==2.26.0
timm==0.4.12
transformers==4.8.2
einops
